add comment

+++ models/pkg.go
@@ -43,11 +43,11 @@ func AssertKeyValueBool(w http.ResponseWriter, m map[string]interface{}, key str
 }
 
 func AssertKeyValueInt(w http.ResponseWriter, m map[string]interface{}, key string) (int, bool) {
-	v, ok := m[key].(int)
+	v, ok := m[key].(float64)
 	if !ok {
 		err := fmt.Errorf("'%s' is required for this request", key)
 		cloudfunc.HttpError(w, err, http.StatusBadRequest)
 		return 0, false
 	}
-	return v, true
+	return int(v), true
 }
